Refactor histogram holder class
Refactoring of the histograms holder class. Remove unnecessary histogram. Simplify the write to file of the histograms. A map of directory name to the list of histogram to be written in such a directory was introduced. The writing to file is now generic. If the directory already exists is also available for others, if not it is create as the name of the key of the map.