Add spdlog library for more flexible message logging
Add the spdlog library to the source code to handle a more
flexible message logging. Instead of normal use of std::cout and
std::cerr, the spdlog library provides a much faster printing out
of messages. It also allows to set level of logging, that can be
then set by command line option.