Change cout print out to spdlog logger
Change as most as possible the std::cout print out of messages to
the spdlog logger. In the logger is saved in Attributes class which
allows other classes that hold a const ref of Attributes to use it.
AnaHist class holds also logger reference since it can not access
to Attributes. The logger is obtained globally by spdlog::get
function and the registered sink name, which is done in main.