Use final / default / delete keywords for derived classes
Use the final, default, delete keywords from c++11 standard to
define the member functions of derived classes. final as override
is used for the virtual function re-implemented in the derived
class. default is used for constructor or destructor in which
nothing special is performed. unnecessary default constructors are
marked delete. Derived classes that not supposed to derived again
are marked with final keywords.