Add possible configuration of analysis run through a file
Add a class for handling the configuration of the analysis run
through a configuration file. The config file is passed to the
program via the command line option -f/--config. The handling of
the parsing of the command line is now take care by the config
class "FullRecoConfig". Boost property tree is used to hold the
configuration options set via the command line and via the config
file. The case of options with units is also possible.
The Makefile is also updated to add the source file.