Update MainSimu with class FullRecoConfig for config handling
Update the main of the analysis code with the FullRecoConfig
class to handle the configuration. The command line parsing is now
inside the FullRecoConfig class. In addition the configuration
through a config file is now possible too. The variables obtained
before through the parsing is now obtained via the Get member
function of the class FullRecoConfig.
Formatting has been also cleaned up, as well as the obsolete
variables and include declarations.