Adjust Histogram handling class to multi-threading
The class that handle the histograms of the analysis is adjusted to
multi-thread paradigm. Mainly the histogram pointers are now a storage
class. The storage class is composed of the initial histogram class
that is used as a definition, a vector of histograms in which is
stored the pointer of the local histogram that is created in each
different threads. Each thread now hold their local histograms class
which is all centralized in histogram handling class in order to
perform the final merge and write procedure to end the analysis.