Adjust FullRecoEvent and FullRecoTask to multithread framework
FullRecoEvent and FullRecoTask are not fundamentally changed for the
multithreading. Those classes are used only on single thread
process. Still the process call from the main and the handle of the
histogram needs to be adjusted. As well PDG_fromName function from
BuildDetectorLayer is moved in FullRecoEvent for more consistency
between headers.