Add return type enum for clarifying information passing in processes
Add the namespace ReturnRes and the return type enum InfoM to clarify
the information on the current status of the process and
analysis. When there is a need to tell to the following step that
something went wrong (BuildError, KalmanError, etc) this enum type
will handle the proper encoding of the information. When the procedure
is going fine, the Fine status is passed through the steps. The
formalization into a enum type simply and clarify the handling of the
steps in the multi-threading framework.