Add new configuration options to the attributes of the analysis
Add the options of configuration needed to distinguish the
multi-thread and the single-thread processing. One is using
MultiThreading and based on blocking queues to process different steps
of the analysis in parallel. The other is using ZeroMQ in order to
obtain the multi-core multi-process paradigm, which use the zeroMQ
library for message passing through different protocols (inproc, ipc,
tpc). All those options must be passed by the config file, that is
parsed and defines the options at run time.