Adapt the main to the different processing models
Adapt the main to the multi-thread processing. Depending whether it is
simple-thread (as initially) or multi-thread (local multi-thread with
blocking queues or zeroMQ multi-core multi-thread) the main function
needed to be adapted to those scenarios. The event loop is still
explicitly call in the single-thread case, however in the both
multi-thread case the event loop is handle within the multi-thread
management. The supervision of the histograms (local or multi-thread)
is transparent in the main function since it is handled by the
histogram class and local storage and central references.