Add blocking queue algorithm to perform the local multi-threading
Add blocking queue algorithm that will handle the asynchronous
processes that correspond to the different step of the analysis. Q1 ->
Q2 -> Q3 : if Q1 is empty Q2 is waiting for following event to
process, same between Q2 and Q3. if Q2 is full, Q1 is waiting before
pushing the event to process into Q2, same between Q3 and Q2. And so
on until the final event is processed. At this moment, special case is
handled to close the previous queue before closing the full analysis.