Add class FullRecoTask to handle multithreading with blocking queue
Add the class FullRecoTaskMT to handle the multithreading and so the
asynchronous processing of the event analysis via the use of blocking
queues. Each intermediate is handled by a blocking queue: after
process the incoming event, the process resulting data structure is
push into a blocking queue (index of the processed event to the data
structure itself). The next step get the following event to process
from the blocking queue (or wait for it). The merger class is used to
merge the different data structure obtained in the various steps into
the ROOT tree. Number of threads for each step are defined by the
configuration of the analysis. The gain of using this process topology
is about 15% to 20% from the single-process style.