Add the MT equivalent of the process steps
Add the MT (multithread) equivalent of the singlethread process steps:
the Builder step for detector planes, the Kalman filter step. One
additional step responsible for the merging the data stream:
MergerOutput. This step is only needed in the context of
multithreading, and will gather the output of the different steps to
be prepared to be written on the ROOT tree and ROOT file.