Add ZeroMQ thread model to handle the multi-core multi-cpu scheme
Those files are responsible for the multithreading via ZeroMQ
framework that will take care of the message communication between the
registered steps. In ThreadModelZMQ is defined the state machine classes
that will be used for handle the abstraction of the tasks. The ZeroMQ
message communication used follow the pull-push paradigm. In
ThreadTasks the mechanism of tasks is defined using the state machine
abstractions defined in ThreadModelZMQ. One task definition for
providing the next event to process. One task definition for N->M
pull-push handling: normally only N -> 1 or 1 -> N pull-push
scheme is possible, to have N -> M  scheme, a proxy is needed
in the middle: N -> 1 -> M. One task definition for builder, Kalman
filter, and merger. One task definition to handle the halt of the code.