Add new RecoEvent and RecoTask classes using the ZeroMQ framework
Add new files FullRecoEvent with the classes defining the data class
used in the message communication via ZeroMQ. Msgpack is used to
perform the binary serialization of the data structure before sending
it via the message queue, and the deserialization after receiving the
message. The msgpack library use macro to define the data structure to
serialize. The files FullRecoTask handle the initialization of the
thread model define in the configuration file and start of the
threads. The thread model will then be self controlled and will end at
the final event and halt message.