Add MainMonitor program to monitor the ZeroMQ setup in realtime
Add a program, MainMonitor, to monitor in realtime the ZeroMQ
multi-core / multi-cpu analysis. The status of the analysis is sent
via http and received by the MainMonitor that set up an http server
provided by ROOT. Then opening any web browser and enter address
of the server and the 8080 port will provide the ROOTjs front-end to
the analysis and histogram that monitor the progress of the analysis.