Add ZeroMQ and Msgpack dependency to the build
Add the ZeroMQ library and msgpack library to the makefile to
represent the new dependencies. The ZeroMQ framework needs two
separate dependency to fulfill: libzmq corresponding to the ZeroMQ
framework in C, and ccpzmq the C++ binding to the C API. The
directories needed will be src/libzmq and src/ccpzmq. The msgpack
library is also needed and placed in src/msgpack-c.