Update README.md with explanation on external libraries needed
Update the README.md with some explanation on the libraries needed.
Eigen3 library is now needed by genfit2. libeigen3-dev needs to
install by the package manager of the system, it simplifies its
management. As well, boost was needed. It was not explicitly said
that boost needs to be from the package manager of the system,
therefore it is now mentioned.