Fix the order Id of the fibers to fill the output branch of Fibers
Fix the order of Ids of the fibers for filling the output tree
branch of the Fibers. The order was defined differently before
the merge with the dev_ekawa_minifiber branch and the order was
keeped was in the dev_ekawa_minifiber, messing up the if condition.
The condition was always false in this case, and needed to be fixed.
Now MiniFiber ids are defined before the one of Fibers, and so if
conditions changed.