Fix Name to order Id in Builder process
Fix the Name to order Id in the BuildDectorLayerPlaneDAF process.
In the merge from dev_ekawa_minifiber additional names (for minifiber)
were included in the Name to OrderId of the detectors. It leaded to
DetId not corresponding to the proper detector branches etc. Leading
to memory leaks and memory corruption.