Add proper treatment of the first hit for path length calculation
Add the proper treatment of the first hit for the path length
calculation. The path length is obtained from the fitter from the
first measurement to the last one. To have the path length total
the distance from target to the first detector is needed and then
added to the obtained path length for proper mass estimation.