Add configurable option for Kalman and CheckRZ step
Add configurable option for Kalman filter and CheckRZ step.
They are obtained from the configuration file and then via the
attributes class in those respective classes.
For Kalman filter, the choice of the fitter is not configurable
without recompilation. The cut for multiplicity in MDC and on the
miniFiber for rejection can also be set from the configuration
file without recompilation.
For CheckRZ the choice of the measurement update can be set via
the configuration, as well the use of the correction of the bias on
the Z estimation.
All of those parameters have default values and the configuration
file is not mandatory to run the analysis.