Add Builder part for extracting hits of PSBE endcap
Add Builder part responsible for extracting and preparing the
genfit hit of the PSBE endcap (in case of being WasaSide=1).
The PSBE bar are represented as planar measurement the local
vertical axe (V) is the radius vector of the radial segment of
the bar. U is orthogonal to V. The Phi angle is within the +- 3.75
degree. Initially, the possible measured radius is within 6. cm
and 22. cm.