Add the RZ fit for the PSBE case and Z estimation in the PSBE case
Add the RZ fit for the PSBE case. First the PSBE does not have any
R measurement, so the miniFiber fit is used to estimate the R position
in the PSBE bar. Then the RZ fit and Z estimations for the MDC layers
is performed as in the case of PSCE. Then the PSBE measurements are
updated with the estimated R position. The radial measurement in the
local plane is the V axis, and U is then R*sin(phi). Phi is uniformly
distributed, so the its variance is Sqrt(2*R*sin(phi_max/2.))/12.
The variance for R is obtained during R estimation. The Bias is not
the same in the PSCE and PSBE case, so an additional new vector
gathered the additional bias to be corrected in the PSBE case.