Add new process to generate an output TTree for ML studies
Add a new process class to generate an output TTree into a new
ROOT file for machine learning studies. The output is related to
the observables that are not directly from the simulation outputs
but are more processed information.
In the current case, it includes the digitized and physics information
at the detector levels, but can be adjusted to the needs of the
ML study. The TTree in this case has to have a flat Ntuple
representation in order to be used either in TMVA or in python
ML frameworks.