Change file, hist and TTree ownership in memory for ML output file
Change of the ownership of the histograms created in the framework to
the memory. They are then written done in the output file of the
reconstruction, including the TTree of the reconstruction. Before
the ownership was implicit and there was only one output file.
Now with the possible presence of the ML output, the ownership
must be clarified to avoid segmentation faults at the end of the
reconstruction.