Encapsulate the data structures for the ML study to be selected
Encapsulate the data structure of the output TTree for the ML study.
Different data structure can now be coded and selected via the
configuration file without to have to recompile the analysis code.
DataML virtual class are use as interface for the different data
structures. New data structure for ML must overload the DataML class.
The constructor method set the branches into the TTree ouptut, and
the FillEvent method is for fill the event tree and must call Fill()
from TTree once the observables are calculated and set to values
attached to the branches. The DataML_Out configuration option
will set the dataML output structure from the config file.