Add data structures of primary vertex reco in FullRecoEvent class
Add additional data structures needed for the primary vertex
reconstruction in the main data structure of the analysis. Si_Hits
is filled by the builder class, BuildDetectorLayerPlaneDAF class.
It represents the channel id of the silicon and its measured energy.
IteractionPoint is used to same the MC information for efficiency
calculation.