Add TRiemannFinder for track finding in solenoid via Riemmann Sphere
Add the TRiemmannFinder class for track finding process in the
solenoid via Riemann Sphere transformation and vis DKTree search of
closest hits. The hits from the MDC and the central barrel and backend
are selected and sorted into track candidates. This is the premise
for the cellular automata algorithm for track finding.
First, MDC hits from parallel wires and PSCE hits are considered for
track candidate seeds, then stereo-wires are added to the track
candidates when the fitted circle of the candidates cross the wire
of interest. Then PSBE hits are considered when the phi projection
of the track candidates overlap on the range in phi of the PSBE
segment of interest.