Change processing of MDC wire from its min to max length of wire
Change the processing the MDC hit of the wire of interest. Before,
the length of the wire to define the direction of the wire were
between -10 to 10 cm over the wire, instead of being its full length.
This led to incorrect local projection onto the wire for genfit
measurement. After the change, the proper length of the wire is
obtained and used to define the genfit measurement.