Finalize the procedure of finding tracks with Riemann sphere
Finalize the procedure of finding tracks with Riemann sphere.
The Hits from PSBE is added. The others hits from stereo wires are
then added. The neighbor hits from MDC or PSCE are check and added
if corresponds to the condition of the track candidate. Then the
position in Z is obtained from the stereo wires and PSCE/PSBE, which
leads to sort the hits by Z position. Finally a helix fit is perform
on the full track candidate via a fit on the Riemann sphere.
Outputs is then prepared into a data structure in RecoEvent for
further steps.