Add define HomogeneousField or NonhomogeneousField to compile flags
Add the define variable HomogeneousField or NonhomogeneousField to
the compile flags to make use of the KFParticle library. Some of
define in the KFParticle are present in their include files, while
the library is already compiled with the corresponding flag. If
the define is passed also during the compilation the KFParticle
headers does not have the define defined and the code is not proper
implemented during the compilation.