Add step FindingPerf to assess performance of the track finders
Add a new step TFindingPerf that is responsible for assessing the
performance of the track finders. It will compare the real tracks
and the found tracks from the track finding algorithms. The real
tracks are rebuilt from the simulation into their hit collections.
Then the track founds are compared to the set of real tracks.
The differences between the found and real tracks is then summarized
into histograms.