Add the WASA field map into the full reconstruction
With the addition of the WASA field map to the simulation, the
analysis code also needs to have the field included. The class MField
used in the Geant4 is copied, with its file extensions changed from
.cc / .hh to .cxx / .h, to fit with the FairBase style.
The Field map class MField is encapsulated into the
WasaSolenoidFieldMap class in order to have common interface via
FairField for genfit to use to selected field. An Eve class is also
prepared for visualization and track propagation via Eve library.
The field map is used in the same way as in Geant4.