Add WasaSolenoidFieldMap as possible field to be used with genfit
Add the WasaSolenoidFieldMap as a possible field to be used.
Normally everything is decided by the input ROOT file that comes
from the Geant4 simulations. In the simParameters saved in the
simulation output file, an additional field will be present when
the field map was used in the simulation. This condition will
determine the selected field for the analysis. As well the maximum
of the field (scaled from the file of the field map) is also saved
and loaded from the Geant4 simulation output.