Add AnaEvent_Metadata class for the output of the analysis run
Add Metadata class, AnaEvent_Metadata, for gathering and saving
metadata of the analysis run. The object will be saved in the output
file side-by-side with the TTree. It will be used to save important
information that will be used for restarting the analysis from a
already finished step. The hash will be the key to the database.
The hash is calculated from the NameIn, NameOut and DataOfRun.