Add save and load function for the metadata class into the output
Add functions to save and load the metadata class AnaEvent_Metadata.
The saving done at the FullRecoTask since it is the Task manager that
owns the Attributes and has the list of tasks. The loading is
performed in the THyphiAttributes which holds the DataSim structure
that gathers the different data attributes from the input file.