Add the read and write of the metadata in the main
Add the read and write of the metadata class in the main. The
read out from the input file is hold by the DataSim InputPar object.
The write into output file is first set by metadata object which is
fill by the task manager. It is then written down into the output
file.