Add the possibility to restart from the previous output TTree file
Add the possibility to restart the analysis from the output TTree
file. The previous analysis saves the metadata with results of the
steps that was selected for this analysis. With the new builder class
TBuildRestarter, the data structure FullRecoEvent is used in the
analysis to gather the intermediates results and data structure. In
order to restart the analysis, the new builder has to reload the
data structure in FullRecoEvent that was written in the TTree file.
Depending the step in which the previous analysis finished, the data
structure needs to be reloaded before being able to continue with the
new steps. The task manager needs a task and a new event loop in which
the input is now a MCAnaEventG4Sol object. Since the two
MCAnaEventG4Sol event object is created and present in the analysis,
the static members of this class needs to be removed. The main is now
able to define either it is a restart or not from the configuration
file.