Add SQLite database for saving the config and attributes of a run
Add a SQLite database for saving the configuration and the
attributes of the run. The sqlite database is inside the
WasaFRSexp.sqlite file. First the database is for book keeping.
Each run is ided by the input file name, the output file name
and the datetime of run. They are hashed to provide unique ID.
It saves the configuration file used saved as JSON format in a
database field. The Attributes of the run (THyphiAttributes class)
are saved via the AttrOut struct.
The idea of the database will be also to save/load the calibration
parameters. For each detector subsystem needing parameters, one table
will be used. Each calibration will be identified by lmd unique
id label (from lmd id to lmd id). When loading at the beginning of
the analysis, the configuration file will define which calibration
to load. Since it will be from configuration, it will be saved in book
keeper table.
In case of the restarting analysis, the metadata from the input file
will be used to find the attributes and configuration that was saved
in the database. The restarted analysis will then have the same
conditions as the previous run, avoiding possible problems around the
use of different parameters.