Update TDataProcess classes to be more generic to output structures
Update the TDataProcess classes in order to be more generic. They
are initially template classes but for the simulation code, they
were fully instantiated. Now with the preparation of the framework
to work with both simulated and experimental data, we need to make
two output data structure similar for its end points, but differ
on the data structure of the detectors data.
Now the code is prepared for the output data structure
MCAnaEventG4Sol. The template class are generic on the output data
structure and full instantiated at the end of their source code to
help the compiler to create the instantiated templates. Once the new
class for the output of the experimental data is available, the
instantiation with such output will be added.