Fix access to histogram object of the base class into the daughters
Fix the access to hthe histogram object AnaHisto that is located in
the base class TDataProcess. With the partial instantiation of the
daughter classes, the compiler could not find the relationship of the
inherited member variable. The use of "this" clarify that it is the
member variable that we want to access to.