Update BuilderRestart to be generic from Simu or Exp In/Out data
Update the BuilderRestart to be generic from the data structure
used to restart the analysis. It is now independent from the
simulation data structure and the experimental data structure.
Depending which one is, the template implementation will be
selected from the type of output was used, and so the input of the
restarter.