Change the database structure to allow g++ to compile the template
Change the database structure to allow g++ to compile the templated
definition of the database. Instead of being in a single table to
gather all the run parameters, several smaller tables are now used.
It allow g++ to compile the THyphiAttributes class. Until now, only
clang++ could finish the compilation. With smaller templated template
definition, g++ can be also used. Be aware that the reloading now
use the new database structure and so the restart of the analysis
from previous database can not perform anymore.