Fix extraction of geometry volumes from the new WASA geometry
Fix the extraction of the geometry volumes needed in the Builder
steps. From the Id of the detector segments, we need to make it
correspond to the index of the nodes in the WASA geometry ROOT file.
The WASA geometry ROOT file has been updated and so we need to be
able to read the previous and new geometry files.