Remove veto of UFT3 extraction from the simulation input
Remove the veto of the UFT3 in the builder of the simulation data.
Before the UFT3 was placed in front of the target, and was not used
for the tracking. In the WASA-FRS experiment the fiber tracker UFT3
was placed behind the target. Therefore, now in the simulation also
has UFT3 behind the target.