Fix name of the mother volume of the minifiber from the new geometry
Fix the name of the mother volume of the minifiber. In the new
geometry the minifiber is seperated into 2 mother volumes, each
having x, u, v layer. Before that, the geometry of the minifiber
had x1u1v1 x2u2v2 layers in one mother volume. So now the names
of the mother volume have to be correctly written.