Include time resolution uncertainty to Simu time information
In SimuBuilder, the time information in the detectors is now
including the effect of their resolution so that now the time
value is randomise within a gaussian distribution. Moreover,
the bug in THypHIAttributes.h that affected the SQL database
and was already fixed in the branch maindata is also fixed now.