Correct the inclusion of PyTorch into the Reco analysis
Correct the inclusion of the PyTorch framework into the reconstruction
analysis. The Makefile has to be set correctly in order to isolate
the PyTorch library and include for only GNNFinder and FullRecoTask.
Also in the TGNNFinder the header that is needed is the torch/torch.h
and not the jit script part of the PyTorch.