Add the global setting for genfit material configuration
Add the global setting for turning off the genfit material behavior
inside of the noise propagation into the Kalman Filter. The option
to disable all material effects is already present with the option
DoNoMaterial. The options DoNoBeth and DoNoMultiScat is now added
in order to have more granularity for the configuration. At the moment
it is global setting, later it will be also only for the tracks that
failed the fitting a first time, where we will turn off the energy
loss calculation in order to have a convergence of the fit.