Fix bug in PerfectFinder which required PSB&PSBE hits
I fix a bug in WasaFinder Task that was requiring both PSB and
PSBE hits when carrying out the perfect finder. Moreover, the
names of PSFE and PSBE are modified to match simudata, where
PSFE is in front and PSBE is in the back of MDC of the real
exp setup, respectively.