Fix path to the GNN model loaded in the torch framework
Fix the path to the GNN model. It was hardcoded to a path that is
not standard. The model is located in the calibration git repository
so it is accessible through it. The name of the standard directory
is "calib", and the default model is given by the path
"calib/gnn_model_nodecl/model_nodecl.pt". So now this path is stored
in the HyphiAttributes class and can be modified via the config file.