Fix bug of missing initialization of Wasa_Side when MainData is used
Fix the bug that Yiming found: when the MainData binary is used for
the experimental data analysis the value of Wasa_Side is not
initialized, which leads to have the signDir of the magnetic field
to be -1.0, while it should be 1.0. In the SimuData binary for the
analysis of the simulated data, the variable Wasa_Side is initialized
from the InputPar.SimParameters. InputPar.SimParameters are loaded from
the ROOT file output of Geant4 simulation.
Now the code will set correctly the value of the Wasa_Side for the MainData
binary.