Add new task to improve Hit of Fiber and PSFE / PSBE
New task that improve the hit representation of the Fiber and
PSFE - PSBE hits. The fibers are represented by 1D hit in genfit, at
the level of the detector, coincidence of layer can be used to improve
the hit representation into 2D hit. In case of PSFE and PSBE, after
the RZ fit from the MDC analysis, the R position in the plastic bar
can be estimated, which allow the hit of PSFE/PSBE to be better
utilized in the Kalman filter fit.